fix : studyRuleController

+++ src/main/java/com/baeker/baeker/studyRule/StudyRuleService.java
@@ -2,8 +2,11 @@ package com.baeker.baeker.studyRule;
 
 import com.baeker.baeker.base.request.Rq;
 import com.baeker.baeker.base.request.RsData;
+import com.baeker.baeker.myStudy.MyStudy;
 import com.baeker.baeker.rule.Rule;
 import com.baeker.baeker.rule.RuleForm;
+import com.baeker.baeker.study.Study;
+import com.baeker.baeker.study.StudyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,16 +27,19 @@ public class StudyRuleService {
 
     private final StudyRuleRepository studyRuleRepository;
 
+    private final StudyService studyService;
+
 
     /**
      * 생성
      */
 
-    public RsData<StudyRule> create(StudyRuleForm studyRuleForm, Rule rule) {
+    public RsData<StudyRule> create(StudyRuleForm studyRuleForm, Rule rule, Study study) {
         StudyRule studyRule = StudyRule.builder()
                 .name(studyRuleForm.getName())
                 .about(studyRuleForm.getAbout())
                 .rule(rule)
+                .study(study)
                 .build();
         studyRule.getRule().getStudyRules().add(studyRule);
         studyRuleRepository.save(studyRule);
@@ -62,10 +68,14 @@ public class StudyRuleService {
      * 삭제
      */
 
-    public void delete(StudyRule studyRule) {
-        studyRule.getRule().getStudyRules().remove(studyRule);
-        studyRuleRepository.delete(studyRule);
-        RsData.of("S-1", "삭제되었습니다");
+    public RsData<StudyRule> delete(StudyRule studyRule, String leader, String user) {
+        if (leader.equals(user)) {
+            studyRule.getRule().getStudyRules().remove(studyRule);
+            studyRuleRepository.delete(studyRule);
+            return RsData.of("S-1", "삭제되었습니다");
+        } else {
+            return RsData.failOf(studyRule);
+        }
     }
 
     /**
@@ -90,8 +100,23 @@ public class StudyRuleService {
 
     public Page<StudyRule> getList(int page) {
         List<Sort.Order> sorts = new ArrayList<>();
-        sorts.add(Sort.Order.desc("createDate"));
+        sorts.add(Sort.Order.desc("selectDate"));
         Pageable pageable = PageRequest.of(page, 10, Sort.by(sorts));
         return studyRuleRepository.findAll(pageable);
+    }
+
+
+    /**
+     * 검증
+     */
+
+    public RsData<Study> isExist(Rq rq, Long id) {
+        RsData<Study> rsData = studyService.getStudy(id);
+        if (rsData.isSuccess()) {
+            if (rsData.getData().getLeader().equals(rq.getMember().getNickName())) {
+                return RsData.of("S-1", "리더 입니다." , rsData.getData());
+            }
+        }
+        return RsData.of("F-1" , "리더가 아닙니다.");
     }
 }
